Split out each submodule into separate files
