+++ python/src/CMakeLists.txt
@@ -2,13 +2,22 @@
 # file: CMakeLists.txt
 # -------------------------------------------------------------
 # -------------------------------------------------------------
+#
+# Copyright (c) 2013 Battelle Memorial Institute
+# Licensed under modified BSD License. A copy of this license can be found
+# in the LICENSE file in the top level directory of this distribution.
+#
 # -------------------------------------------------------------
 # -------------------------------------------------------------
 # Created February 17, 2020 by Perkins
-# Last Change: 2023-11-30 09:33:03 d3g096
+# Last Change: 2025-02-10 12:44:07 d3g096
 # -------------------------------------------------------------
 
-pybind11_add_module(gridpack gridpack.cpp)
+pybind11_add_module(gridpack
+  gridpack.cpp
+  gridpack_ds.cpp
+  gridpack_emt.cpp
+  )
 
 set(parallel_scripts 
   ${CMAKE_CURRENT_SOURCE_DIR}/hello.py
